Reconfigure Checkstyle plugin to remove absolute path inputs
As configured before, the `configProperties` field has an absolute path
sensitivity introduced by the reference to the suppressions file. This
change causes the path to be relative to the project root, which can
facilitate cache hits when absolute paths change (for example, when the
build is run from a CI system and a remote cache is populated, then consumed
from a local build).

This refactoring matches recommendations found in the Gradle documentation for
structuring Checkstyle configuration files:

https://docs.gradle.org/current/userguide/checkstyle_plugin.html#sec:checkstyle_built_in_variables

This change also configure Checkstyle via the `CheckstyleExtension` rather than
the `Checkstyle` tasks